rpc_report/requirements.txt
grpcio==1.76.0
grpcio-tools==1.76.0
python-dotenv==1.2.1
pandas==2.3.3
openpyxl==3.1.5
python-docx==1.2.0
fpdf2==2.8.5
matplotlib==3.10.7